@@ -1,19 +1,19 @@
 name: update_available_android
 description: Android platform implementation of update_available
-version: 1.0.1+1
+version: 2.0.0-nullsafety
 repository: https://github.com/mateusfccp/update_available/tree/master/update_available_android
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  update_available_platform_interface: ^1.0.1
+  update_available_platform_interface: ^2.0.0-nullsafety
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  pedantic: ^1.11.0
 
 flutter:
   plugin:
