@@ -1,22 +1,22 @@
 name: update_available_ios
 description: iOS platform implementation of update_available
-version: 1.0.1+1d
+version: 2.0.0-nullsafety
 repository: https://github.com/mateusfccp/update_available/tree/master/update_available_ios
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.2
-  meta: ^1.1.8
-  package_info: '>=0.4.3 <2.0.0'
-  update_available_platform_interface: ^1.0.1
-  version: ^1.2.0
+  http: ^0.13.0
+  meta: ^1.3.0
+  package_info: ^2.0.0
+  update_available_platform_interface: ^2.0.0-nullsafety
+  pub_semver: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.9.0
+  pedantic: ^1.11.0
