@@ -1,14 +1,14 @@
 name: update_available_platform_interface
 description: A common platform interface for the update_available plugin.
-version: 1.0.1+1b
+version: 2.0.0-nullsafety
 repository: https://github.com/mateusfccp/update_available/tree/master/update_available_platform_interface
 
 environment:
-  sdk: '>=2.9.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  meta: ^1.1.8
-  plugin_platform_interface: ^1.0.2
+  meta: ^1.3.0
+  plugin_platform_interface: ^2.0.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  pedantic: ^1.11.0
