@@ -1,2 +1,2 @@
-String Function(String json) getStringByKey(String json) =>
+String? Function(String json) getStringByKey(String json) =>
     (String key) => RegExp('"$key":\\s*"([^"]*)"').firstMatch(json)?.group(1);
